requirements.txt
Flask
pymongo
Pillow
numpy
torch
torchvision
opencv-python
mediapipe
onnxruntime